requirements.txt
croniter==1.3.7
types-croniter==1.3.2
hikari==2.0.0.dev111
hikari-lightbulb==2.2.4
python-dotenv==0.21.0
